🌳 Refactor: email 관련 api 분리
